+++ src/vaster.h
@@ -1,0 +1,8 @@
+void check_range(SEXP);
+void check_size(SEXP);
+
+void check_extent(SEXP);
+void check_dimension(SEXP);
+SEXP row_from_y_(SEXP, SEXP, SEXP);
+SEXP col_from_x_(SEXP, SEXP, SEXP);
+SEXP bin_from_float(SEXP, SEXP, SEXP);