+++ src/coordinates.c
@@ -42,3 +42,31 @@ SEXP row_from_y_(SEXP nrow, SEXP ylim, SEXP py)
   return bin_from_float(nrow, ylim, py);
 
 }
+
+SEXP coord_from_bin(SEXP bins, SEXP range, SEXP index) {
+  int nn = LENGTH(index);
+  SEXP out;
+  out = PROTECT(Rf_allocVector(REALSXP, nn));
+
+  int nbins = INTEGER(bins)[0];
+  double res = (REAL(range)[1] - REAL(range)[0])/nbins;
+  double cmin = REAL(range)[0];
+  double* rout = REAL(out);
+  int* rindex = INTEGER(index);
+  for (int i = 0; i < nn; i++) {
+    if ((rindex[i] >= 0) && (rindex[i] < nbins)) {
+      rout[i] = cmin + res/2 + rindex[i] * res;
+    } else {
+      rout[i] = R_NaReal;
+    }
+  }
+  UNPROTECT(1);
+  return out;
+}
+SEXP x_from_col_(SEXP ncol, SEXP xlim, SEXP col) {
+  return coord_from_bin(ncol, xlim, col);
+}
+
+SEXP y_from_row_(SEXP nrow, SEXP ylim, SEXP row) {
+  return coord_from_bin(nrow, ylim, row);
+}