+++ R/vrt_extent.R
@@ -1,19 +1,24 @@
-#' Extents from VRT
+#' Extents from VRT (virtual raster data set of GDAL)
 #'
-#' Get extent from index values in VRT text.
+#' Get extent from index values in VRT text, these are the individual footprints of raster
+#' windows in a VRT. These can be arbitrary to a specific grid, but generally are used for tiled mosaics.
 #'
-#' (I can't understand XML tech  so I hack the text as lines with strsplit)
+#' Each VRT raster element records it's relative position within the grid, so we grid logic
+#' to expand the actual extent of each element, and return those as a matrix of xmin,xmax,ymin,ymax.
+#'
+#'
 #' @param x url or file path to VRT file
 #'
+#' @returns a matrix of 4 columns, xmin,xmax,ymin,ymax
 #' @export
 #' @examples
 #' #src <- "https://opentopography.s3.sdsc.edu/raster/NASADEM/NASADEM_be.vrt"
-#' #src <- "https://opentopography.s3.sdsc.edu/raster/SRTM_GL1/SRTM_GL1_srtm.vrt"
-#' #ex <- extent_vrt(src)
-#' #op <- par(mar = rep(0, 4))
-#' #plot(range(ex[,1:2]), range(ex[,3:4]), xlab = "", ylab = "", asp = "", type = "n")
-#' #rect(ex[,1], ex[,3], ex[, 2], ex[,4])
-#' #par(op)
+#' src <- gzfile(system.file("extdata/NASADEM_be.vrt.gz", package = "vaster"), "rt")
+#' ## read VRT from a URL or file (we use a connection here to keep package example small)
+#' l <- base::readLines(src)
+#' ex <- extent_vrt(src)
+#' close(src)
+#' plot_extent(ex)
 extent_vrt <- function(x) {
   vrt <- readLines(x)
 
@@ -40,7 +45,3 @@ extent_vrt <- function(x) {
   cbind(xmin = xmin, xmax = xmax, ymin = ymin, ymax = ymax)
 }
 
-
-show_tiling <- function(overviews, extent) {
-
-}